Change mapping of uint{8,16,32,64}
This changes the mapping of uint types to non-widening yang.common
classes and deals with them being used by yang.{data,model}.api:

YANG type  Old Java type         New Java Type
uint8      java.lang.Short       org.opendaylight.yangtools.yang.common.Uint8
uint16     java.lang.Integer     org.opendaylight.yangtools.yang.common.Uint16
uint32     java.lang.Long        org.opendaylight.yangtools.yang.common.Uint32
uint64     java.math.BigInteger  org.opendaylight.yangtools.yang.common.Uint64

This affects all getters, setters and value wrapper classes. For
setters, code generator emits a deprecated setter, which performs
the conversion internally.

JIRA: MDSAL-330
JIRA: YANGTOOLS-1018

+++ binding/mdsal-binding-generator-impl/src/test/java/org/opendaylight/mdsal/binding/yang/types/BaseYangTypesTest.java
@@ -12,7 +12,6 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.util.Set;
 import org.junit.BeforeClass;
 import org.junit.Test;
@@ -20,6 +19,10 @@ import org.opendaylight.mdsal.binding.generator.spi.TypeProvider;
 import org.opendaylight.mdsal.binding.model.api.Type;
 import org.opendaylight.mdsal.binding.model.util.BindingGeneratorUtil;
 import org.opendaylight.yangtools.yang.common.Empty;
+import org.opendaylight.yangtools.yang.common.Uint16;
+import org.opendaylight.yangtools.yang.common.Uint32;
+import org.opendaylight.yangtools.yang.common.Uint64;
+import org.opendaylight.yangtools.yang.common.Uint8;
 import org.opendaylight.yangtools.yang.model.api.SchemaContext;
 import org.opendaylight.yangtools.yang.model.api.TypeDefinition;
 import org.opendaylight.yangtools.yang.model.api.type.BinaryTypeDefinition;
@@ -166,19 +169,19 @@ public class BaseYangTypesTest {
 
         javaType = typeProvider.javaTypeForSchemaDefinitionType(uint8, uint8);
         assertNotNull(javaType);
-        assertEquals(Short.class.getCanonicalName(), javaType.getFullyQualifiedName());
+        assertEquals(Uint8.class.getCanonicalName(), javaType.getFullyQualifiedName());
 
         javaType = typeProvider.javaTypeForSchemaDefinitionType(uint16, uint16);
         assertNotNull(javaType);
-        assertEquals(Integer.class.getCanonicalName(), javaType.getFullyQualifiedName());
+        assertEquals(Uint16.class.getCanonicalName(), javaType.getFullyQualifiedName());
 
         javaType = typeProvider.javaTypeForSchemaDefinitionType(uint32, uint32);
         assertNotNull(javaType);
-        assertEquals(Long.class.getCanonicalName(), javaType.getFullyQualifiedName());
+        assertEquals(Uint32.class.getCanonicalName(), javaType.getFullyQualifiedName());
 
         javaType = typeProvider.javaTypeForSchemaDefinitionType(uint64, uint64);
         assertNotNull(javaType);
-        assertEquals(BigInteger.class.getCanonicalName(), javaType.getFullyQualifiedName());
+        assertEquals(Uint64.class.getCanonicalName(), javaType.getFullyQualifiedName());
 
         javaType = typeProvider.javaTypeForSchemaDefinitionType(union, union);
         assertNotNull(javaType);
@@ -239,22 +242,22 @@ public class BaseYangTypesTest {
         javaType = typeProvider.javaTypeForSchemaDefinitionType(uint8, uint8,
             BindingGeneratorUtil.getRestrictions(uint8));
         assertNotNull(javaType);
-        assertEquals(Short.class.getCanonicalName(), javaType.getFullyQualifiedName());
+        assertEquals(Uint8.class.getCanonicalName(), javaType.getFullyQualifiedName());
 
         javaType = typeProvider.javaTypeForSchemaDefinitionType(uint16, uint16,
             BindingGeneratorUtil.getRestrictions(uint16));
         assertNotNull(javaType);
-        assertEquals(Integer.class.getCanonicalName(), javaType.getFullyQualifiedName());
+        assertEquals(Uint16.class.getCanonicalName(), javaType.getFullyQualifiedName());
 
         javaType = typeProvider.javaTypeForSchemaDefinitionType(uint32, uint32,
             BindingGeneratorUtil.getRestrictions(uint32));
         assertNotNull(javaType);
-        assertEquals(Long.class.getCanonicalName(), javaType.getFullyQualifiedName());
+        assertEquals(Uint32.class.getCanonicalName(), javaType.getFullyQualifiedName());
 
         javaType = typeProvider.javaTypeForSchemaDefinitionType(uint64, uint64,
             BindingGeneratorUtil.getRestrictions(uint64));
         assertNotNull(javaType);
-        assertEquals(BigInteger.class.getCanonicalName(), javaType.getFullyQualifiedName());
+        assertEquals(Uint64.class.getCanonicalName(), javaType.getFullyQualifiedName());
 
         javaType = typeProvider.javaTypeForSchemaDefinitionType(union, union,
             BindingGeneratorUtil.getRestrictions(union));
